refactor(src): split state/queue.rs by responsibility
